@@ -1,39 +1,38 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace MSTest.Project
+namespace MSTest.Project;
+
+[TestClass]
+public class ByAlphabeticalOrder
 {
-    [TestClass]
-    public class ByAlphabeticalOrder
+    public static bool Test1Called;
+    public static bool Test2Called;
+    public static bool Test3Called;
+
+    [TestMethod]
+    public void Test2()
     {
-        public static bool Test1Called;
-        public static bool Test2Called;
-        public static bool Test3Called;
+        Test2Called = true;
 
-        [TestMethod]
-        public void Test2()
-        {
-            Test2Called = true;
+        Assert.IsTrue(Test1Called);
+        Assert.IsFalse(Test3Called);
+    }
 
-            Assert.IsTrue(Test1Called);
-            Assert.IsFalse(Test3Called);
-        }
+    [TestMethod]
+    public void Test1()
+    {
+        Test1Called = true;
 
-        [TestMethod]
-        public void Test1()
-        {
-            Test1Called = true;
+        Assert.IsFalse(Test2Called);
+        Assert.IsFalse(Test3Called);
+    }
 
-            Assert.IsFalse(Test2Called);
-            Assert.IsFalse(Test3Called);
-        }
+    [TestMethod]
+    public void Test3()
+    {
+        Test3Called = true;
 
-        [TestMethod]
-        public void Test3()
-        {
-            Test3Called = true;
-
-            Assert.IsTrue(Test1Called);
-            Assert.IsTrue(Test2Called);
-        }
+        Assert.IsTrue(Test1Called);
+        Assert.IsTrue(Test2Called);
     }
 }
