@@ -1,52 +1,51 @@
 ﻿using NUnit.Framework;
 
-namespace NUnit.Project
+namespace NUnit.Project;
+
+public class ByOrder
 {
-    public class ByOrder
+    public static bool Test1Called;
+    public static bool Test2ACalled;
+    public static bool Test2BCalled;
+    public static bool Test3Called;
+
+    [Test, Order(5)]
+    public void Test1()
     {
-        public static bool Test1Called;
-        public static bool Test2ACalled;
-        public static bool Test2BCalled;
-        public static bool Test3Called;
+        Test1Called = true;
 
-        [Test, Order(5)]
-        public void Test1()
-        {
-            Test1Called = true;
+        Assert.IsFalse(Test2ACalled);
+        Assert.IsTrue(Test2BCalled);
+        Assert.IsTrue(Test3Called);
+    }
 
-            Assert.IsFalse(Test2ACalled);
-            Assert.IsTrue(Test2BCalled);
-            Assert.IsTrue(Test3Called);
-        }
+    [Test, Order(0)]
+    public void Test2B()
+    {
+        Test2BCalled = true;
 
-        [Test, Order(0)]
-        public void Test2B()
-        {
-            Test2BCalled = true;
+        Assert.IsFalse(Test1Called);
+        Assert.IsFalse(Test2ACalled);
+        Assert.IsTrue(Test3Called);
+    }
 
-            Assert.IsFalse(Test1Called);
-            Assert.IsFalse(Test2ACalled);
-            Assert.IsTrue(Test3Called);
-        }
+    [Test]
+    public void Test2A()
+    {
+        Test2ACalled = true;
 
-        [Test]
-        public void Test2A()
-        {
-            Test2ACalled = true;
+        Assert.IsTrue(Test1Called);
+        Assert.IsTrue(Test2BCalled);
+        Assert.IsTrue(Test3Called);
+    }
 
-            Assert.IsTrue(Test1Called);
-            Assert.IsTrue(Test2BCalled);
-            Assert.IsTrue(Test3Called);
-        }
+    [Test, Order(-5)]
+    public void Test3()
+    {
+        Test3Called = true;
 
-        [Test, Order(-5)]
-        public void Test3()
-        {
-            Test3Called = true;
-
-            Assert.IsFalse(Test1Called);
-            Assert.IsFalse(Test2ACalled);
-            Assert.IsFalse(Test2BCalled);
-        }
+        Assert.IsFalse(Test1Called);
+        Assert.IsFalse(Test2ACalled);
+        Assert.IsFalse(Test2BCalled);
     }
 }
