@@ -1,14 +1,11 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using Xunit.Abstractions;
+﻿using Xunit.Abstractions;
 using Xunit.Sdk;
 
-namespace XUnit.Project.Orderers
+namespace XUnit.Project.Orderers;
+
+public class AlphabeticalOrderer : ITestCaseOrderer
 {
-    public class AlphabeticalOrderer : ITestCaseOrderer
-    {
-        public IEnumerable<TTestCase> OrderTestCases<TTestCase>(
-            IEnumerable<TTestCase> testCases) where TTestCase : ITestCase =>
-            testCases.OrderBy(testCase => testCase.TestMethod.Method.Name);
-    }
+    public IEnumerable<TTestCase> OrderTestCases<TTestCase>(
+        IEnumerable<TTestCase> testCases) where TTestCase : ITestCase =>
+        testCases.OrderBy(testCase => testCase.TestMethod.Method.Name);
 }
