@@ -1,12 +1,9 @@
-﻿using System;
+﻿namespace XUnit.Project.Attributes;
 
-namespace XUnit.Project.Attributes
+[AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
+public class TestPriorityAttribute : Attribute
 {
-    [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
-    public class TestPriorityAttribute : Attribute
-    {
-        public int Priority { get; private set; }
+    public int Priority { get; private set; }
 
-        public TestPriorityAttribute(int priority) => Priority = priority;
-    }
+    public TestPriorityAttribute(int priority) => Priority = priority;
 }
